fix(roas/merge): review polish
- Hoist `record_kept_base_or_error` from each version's `merge.rs`
  into `common/merge.rs`. The helper is version-agnostic and was
  duplicated verbatim four times.
- Move v3.0 `ServerVariable` merge impl from `v3_0/merge.rs` to
  `v3_0/server.rs`, matching the v3.1 / v3.2 convention.
- v2: use `ScalarOverridden` (not `ParameterVariantMismatch`) for
  same-variant Path/Query/Header/FormData leaf-replace and for
  whole-`Header` / whole-`SecurityScheme` leaf-replace. `ParameterVariantMismatch`
  is now reserved for the documented "cross-location" case
  (`Parameter::Path × Parameter::Query` etc.), matching the doc and
  staying consistent with the Body / Items arms that already used
  `ScalarOverridden`.
- v2: `leaf_replace_in_*` take `&str` instead of unused `&mut str`.
- v2: factor `in_path_name` / `in_query_name` / `in_header_name` /
  `in_form_data_name` per-variant helpers so `parameter_ref_key` no
  longer re-matches the outer `Parameter` enum a second time after
  proving the variant.
- v2: `Schema::merge_with_context` matches on owned `other` up front
  in the deep-merge arm, dropping the `let _ = other_obj;` placeholder
  + `unreachable!()` guard.

Verified: 1655 tests pass across all features, clippy clean.
